Auto Go Modules/Go Path mode
In Go 1.16, the default behavior for how Go handles modules has changed.
It now defaults to the Go Modules mode (as opposed to the GO Path mode).
Since we have apps not yet following the go modules mode, we should set
the GO111MODULE environment variable to auto as this will restore the
old behavior where it will switch based on the presence of a go.mod
file.  Check out https://blog.golang.org/go116-module-changes for more
details.